Update README.md to warn users to use the correct Python Version
I have updated the README.md to warn the reader that a specific range of python versions only work for this codebase. I have also included the command to check which Python version you are using. I added a reason why a dedicated virtual environment is needed, and also updated this instruction for installation of this codebase